Created Three buttons for use in the newly added Main Menu scene
A new Scene was created called Main Menu. This is where our menu buttons will be displayed.

Button Manager script was written to manage the actions taken with each respective buton press.

The Button Manager Object was created (with the Button Manager Script added as a component) to use on each button.

A canvas was made with three (3) buttons, each with their own respective text (Credits, Play, Quit). These buttons were then oriented appropriately in the scene.

The Button Manager Object was assigned to "On Click" for each button, and the function respective to the button was assigned according. (ie, LoadSceneCredits() on the Credits button, LoadScenePlay() on the Play button, etc)

The quit button will quit the application, but if the Unity editor is running, it will close the editor instead.

A new scene was Created called "Play". This is the destination for the "Play"