Update PlayerShooting.cs
I use a GameObject array to track the arrows in the scene and change their rotation to the direction that the velocity vector is pointing. If the velocity's magnitude is too low, the arrows begin to spin out of control. I have yet to come up with a good remedy to this, but currently have an if statement in place to sort of fix this