Create RUNE_2.TTF
this is the additional dwarven text titled RUNE_2...its a ttf file. In a dwarven language, not english text. The other font that I already have in the same folder is under the title DWARVESC also a TTF file. Thanks, I did this all through the UI branch and is located as such.  Assets > Textures > UI.
Thanks!